+++ spec/norma_cql_spec.rb
@@ -12,9 +12,9 @@ require 'activefacts/generate/cql'
 include ActiveFacts
 
 describe "Norma Loader" do
-  ORM_CQL_FAILURES = %w{
-    OddIdentifier
-    ServiceDirector
+  orm_cql_failures = {
+    "OddIdentifier" => "Strange identification pattern is incorrectly verbalised to CQL",
+    "ServiceDirector" => "ORM file has incorrect role names, not corrected in order to recall this modeling error"
   }
   # Generate and return the CQL for the given vocabulary
   def cql(vocabulary)
@@ -32,16 +32,23 @@ describe "Norma Loader" do
     actual_file = orm_file.sub(%r{examples/norma/(.*).orm\Z}, 'spec/actual/\1.cql')
 
     it "should load #{orm_file} and dump CQL matching #{expected_file}" do
-      pending if ORM_CQL_FAILURES.include? File.basename(orm_file, ".orm")
       vocabulary = ActiveFacts::Input::ORM.readfile(orm_file)
 
       cql = cql(vocabulary)
       # Save the actual file:
       File.open(actual_file, "w") { |f| f.write cql*"" }
 
-      pending unless File.exists? expected_file
-      cql.should == File.open(expected_file) {|f| f.readlines}
-      File.delete(actual_file)
+      pending("expected output file #{expected_file} not found") unless File.exists? expected_file
+
+      broken = orm_cql_failures[File.basename(orm_file, ".orm")]
+      if broken
+        pending(broken) {
+          cql.should == File.open(expected_file) {|f| f.readlines}
+        }
+      else
+        cql.should == File.open(expected_file) {|f| f.readlines}
+        File.delete(actual_file)
+      end
     end
   end
 end
